fix(inlines): address PR review comments
- Default instruction() on InlineSpec using Self::OPCODE/FUNCT3/FUNCT7
  from InlineOp, remove redundant overrides in all spec.rs files
- Rename message_words to w in blake2 spec reference impl
- Delete blake3 test_utils.rs, inline helpers into test modules
