+++ jolt-inlines/blake2/src/spec.rs
@@ -3,7 +3,7 @@ use crate::{IV, SIGMA};
 use jolt_inlines_sdk::host::Xlen;
 use jolt_inlines_sdk::spec::rand::rngs::StdRng;
 use jolt_inlines_sdk::spec::rand::Rng;
-use jolt_inlines_sdk::spec::{InlineMemoryLayout, InlineSpec, InlineTestHarness, INLINE};
+use jolt_inlines_sdk::spec::{InlineMemoryLayout, InlineSpec, InlineTestHarness};
 
 fn g(v: &mut [u64; 16], a: usize, b: usize, c: usize, d: usize, x: u64, y: u64) {
     v[a] = v[a].wrapping_add(v[b]).wrapping_add(x);
@@ -28,91 +28,27 @@ impl InlineSpec for Blake2bCompression {
     }
 
     fn reference(input: &Self::Input) -> Self::Output {
-        let (state, message_words) = input;
+        let (state, w) = input;
         let mut v = [0u64; 16];
         v[0..8].copy_from_slice(state);
         v[8..16].copy_from_slice(&IV);
 
-        v[12] ^= message_words[16];
+        v[12] ^= w[16];
 
-        if message_words[17] != 0 {
+        if w[17] != 0 {
             v[14] = !v[14];
         }
 
         for s in SIGMA {
-            g(
-                &mut v,
-                0,
-                4,
-                8,
-                12,
-                message_words[s[0]],
-                message_words[s[1]],
-            );
-            g(
-                &mut v,
-                1,
-                5,
-                9,
-                13,
-                message_words[s[2]],
-                message_words[s[3]],
-            );
-            g(
-                &mut v,
-                2,
-                6,
-                10,
-                14,
-                message_words[s[4]],
-                message_words[s[5]],
-            );
-            g(
-                &mut v,
-                3,
-                7,
-                11,
-                15,
-                message_words[s[6]],
-                message_words[s[7]],
-            );
+            g(&mut v, 0, 4, 8, 12, w[s[0]], w[s[1]]);
+            g(&mut v, 1, 5, 9, 13, w[s[2]], w[s[3]]);
+            g(&mut v, 2, 6, 10, 14, w[s[4]], w[s[5]]);
+            g(&mut v, 3, 7, 11, 15, w[s[6]], w[s[7]]);
 
-            g(
-                &mut v,
-                0,
-                5,
-                10,
-                15,
-                message_words[s[8]],
-                message_words[s[9]],
-            );
-            g(
-                &mut v,
-                1,
-                6,
-                11,
-                12,
-                message_words[s[10]],
-                message_words[s[11]],
-            );
-            g(
-                &mut v,
-                2,
-                7,
-                8,
-                13,
-                message_words[s[12]],
-                message_words[s[13]],
-            );
-            g(
-                &mut v,
-                3,
-                4,
-                9,
-                14,
-                message_words[s[14]],
-                message_words[s[15]],
-            );
+            g(&mut v, 0, 5, 10, 15, w[s[8]], w[s[9]]);
+            g(&mut v, 1, 6, 11, 12, w[s[10]], w[s[11]]);
+            g(&mut v, 2, 7, 8, 13, w[s[12]], w[s[13]]);
+            g(&mut v, 3, 4, 9, 14, w[s[14]], w[s[15]]);
         }
 
         let mut result = *state;
@@ -125,14 +61,6 @@ impl InlineSpec for Blake2bCompression {
     fn create_harness() -> InlineTestHarness {
         let layout = InlineMemoryLayout::single_input(144, 64);
         InlineTestHarness::new(layout, Xlen::Bit64)
-    }
-
-    fn instruction() -> INLINE {
-        InlineTestHarness::create_default_instruction(
-            crate::INLINE_OPCODE,
-            crate::BLAKE2_FUNCT3,
-            crate::BLAKE2_FUNCT7,
-        )
     }
 
     fn load(harness: &mut InlineTestHarness, input: &Self::Input) {
